requirements.txt
Flask==3.0.3
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.31.0
langdetect==1.0.9
